requirement.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2021.10.8
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.0
Django==4.0.2
django-celery-beat==2.6.0
django-cors-headers==3.13.0
django-filter==21.1
django-timezone-field==7.0
djangorestframework==3.13.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.1
incremental==22.10.0
isodate==0.6.1
kombu==5.3.7
lxml==5.1.0
Markdown==3.3.6
msgpack==1.0.7
packaging==23.1
Pillow==9.3.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2021.3
redis==5.0.1
requests==2.27.1
requests-file==2.0.0
requests-toolbelt==1.0.0
service-identity==24.1.0
six==1.16.0
smsir==1.0.5
sqlparse==0.4.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.9
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
zeep==4.2.1
zipp==3.7.0
zope.interface==6.2
